BREAKING: session.close() introduced and mandatory
This version supports FOLIO's new (cookie-based) authentication scheme
as well as the old (`X-Okapi-Token`-based) scheme, using the former
when available and falling back to the latter if running against an
older version of FOLIO that does not provide it.

When cookie-based authentication is used, the access token is
periodically refreshed -- when half of its TTL has elapsed. The thread
that does this will not shut down (and so the program will not
terminate) unless `session.close` is called when the session's work is
done.

The documentation and sample program have been updated accordingly.